update a code measuretestfactory
