+++ langtest/transform/__init__.py
@@ -9,6 +9,8 @@ import nest_asyncio
 import pandas as pd
 from tqdm.asyncio import tqdm
 
+from langtest.transform.measure import BaseMeasure
+
 from .accuracy import BaseAccuracy
 from .bias import BaseBias
 from .custom_bias import add_custom_data
@@ -1271,6 +1273,62 @@ class ToxicityTestFactory(ITests):
         tests = {
             j: i
             for i in BaseToxicity.__subclasses__()
+            for j in (i.alias_name if isinstance(i.alias_name, list) else [i.alias_name])
+        }
+        return tests
+
+class MeasureTestFactory(ITests):
+    """Factory class for the robustness measure.
+
+    This class implements the robustness measure. The robustness measure is the number of test cases that the model fails to run on.
+
+    """
+
+    alias_name = "measure"
+
+    @staticmethod
+    def transform(params: dict, *args, **kwargs) -> List[Sample]:
+        """Transforms the sample data based on the implemented tests measure.
+
+        Args:
+            sample (Sample): The input data to be transformed.
+            **kwargs: Additional arguments to be passed to the tests measure.
+
+        Returns:
+            Sample: The transformed data based on the implemented
+            tests measure.
+
+        """
+        pass
+
+    @staticmethod
+    async def run(
+        sample_list: List[Sample], model: ModelFactory, **kwargs
+    ) -> List[Sample]:
+        """Runs the robustness measure.
+
+        Args:
+            sample_list (List[Sample]): The input data to be transformed.
+            model (ModelFactory): The model to be used for evaluation.
+            **kwargs: Additional arguments to be passed to the robustness measure.
+
+        Returns:
+            List[Sample]: The transformed data based on the implemented robustness measure.
+
+        """
+        pass
+        
+    @classmethod
+    def available_tests(cls) -> Dict[str, str]:
+        """Returns the available robustness measure.
+
+        Returns:
+            Dict[str, str]: The available robustness measure.
+
+        """
+        tests = {
+            j: i
+            for i in BaseMeasure.__subclasses__()
             for j in (i.alias_name if isinstance(i.alias_name, list) else [i.alias_name])
         }
         return tests
